Backend/requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
django-extensions==3.0.6
psycopg2-binary==2.9.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
Cerberus==1.3.2
PyJWT==1.7.1
huey==2.4.3
redis==4.2.2
